# backend/openapi.yaml
openapi: 3.0.3
info:
  title: Nimpass REST API
  version: 1.0.0
  license: { name: MIT, url: 'https://spdx.org/licenses/MIT.html' }
  description: |
    All money values are integer Luna. Login and payout proofs must be produced
    by Nimiq Pay's sign(message), using the exact server-provided challenge message.
    Payout completion requires signatures by both the new payout wallet and the
    authenticated owner wallet over that same challenge message.
    Unsafe methods require an exact configured Origin (or Referer fallback).
    Shared middleware errors use the Error schema: ORIGIN_FORBIDDEN (403),
    AUTH_REQUIRED (401), CSRF_INVALID (403), CONTENT_TYPE (415),
    VALIDATION_ERROR (400), RATE_LIMITED (429) and INTERNAL_ERROR (500).
    Session lookup infrastructure failures return INTERNAL_ERROR (500), preserving
    the cookie; AUTH_REQUIRED (401) means a missing, invalid, expired or revoked session.
    Authenticated operations require a session; every unsafe authenticated
    operation requires CSRF. JSON bodies reject unknown fields and trailing data.
    Text length bounds are UTF-8 byte limits after documented normalization.
    Authenticated unsafe methods additionally require X-CSRF-Token, obtained
    from login or GET /auth/session. Cookies are HttpOnly, SameSite=Lax; HTTPS
    deployments use Secure __Host-nimpass_session, local HTTP development uses
    nimpass_session. Requests must send credentials: include.
...
    unchanged.
servers:
  - url: /api/v1
security:
  - sessionCookie: []
paths:
...
  /providers:
    get:
      summary: List own providers
      operationId: listOwnProviders
      responses:
        default: { $ref: '#/components/responses/Error' }
        '200': { description: Own providers, content: { application/json: { schema: { type: object, required: [items], properties: { items: { type: array, items: { $ref: '#/components/schemas/Provider' } } } } } } }
        '401': { $ref: '#/components/responses/Error' }
    post:
      summary: Create provider profile
      description: >-
        The payout wallet is adopted from the authenticated session and the
        provider is returned already verified (DECISIONS.md ADR-025). No field
        of ProviderInput names a payout address; changing one afterwards goes
        through /providers/{providerID}/payout-challenges.
      operationId: createProvider
      parameters: [ { $ref: '#/components/parameters/Csrf' } ]
      requestBody:
        required: true
        content: { application/json: { schema: { $ref: '#/components/schemas/ProviderInput' } } }
      responses:
        default: { $ref: '#/components/responses/Error' }
        '201': { description: Created provider, content: { application/json: { schema: { $ref: '#/components/schemas/Provider' } } } }
        '400': { $ref: '#/components/responses/Error' }
        '401': { $ref: '#/components/responses/Error' }
        '403': { $ref: '#/components/responses/Error' }
        '429': { $ref: '#/components/responses/Error' }
  /providers/{providerID}:
    parameters: [ { $ref: '#/components/parameters/ProviderID' } ]
    get:
      summary: Get own provider
      operationId: getOwnProvider
      responses:
        default: { $ref: '#/components/responses/Error' }
        '200': { description: Own provider, content: { application/json: { schema: { $ref: '#/components/schemas/Provider' } } } }
        '404': { $ref: '#/components/responses/Error' }
    patch:
      summary: Update provider profile
      operationId: updateProvider
      parameters: [ { $ref: '#/components/parameters/Csrf' } ]
      requestBody:
        required: true
        content: { application/json: { schema: { $ref: '#/components/schemas/ProviderInput' } } }
      responses:
        default: { $ref: '#/components/responses/Error' }
        '200': { description: Updated profile; payout fields cannot be mass-assigned, content: { application/json: { schema: { $ref: '#/components/schemas/Provider' } } } }
        '404': { $ref: '#/components/responses/Error' }
        '429': { $ref: '#/components/responses/Error' }
...
  /providers/{providerID}/services:
    parameters: [ { $ref: '#/components/parameters/ProviderID' } ]
    get:
      summary: List own services
      operationId: listOwnServices
      responses:
        default: { $ref: '#/components/responses/Error' }
        '200': { description: Services, content: { application/json: { schema: { type: object, required: [items], properties: { items: { type: array, items: { $ref: '#/components/schemas/Service' } } } } } } }
        '404': { $ref: '#/components/responses/Error' }
    post:
      summary: Create service
      operationId: createService
      parameters: [ { $ref: '#/components/parameters/Csrf' } ]
      requestBody:
        required: true
        content: { application/json: { schema: { $ref: '#/components/schemas/ServiceInput' } } }
      responses:
        default: { $ref: '#/components/responses/Error' }
        '201': { description: Draft service, content: { application/json: { schema: { $ref: '#/components/schemas/Service' } } } }
        '400': { $ref: '#/components/responses/Error' }
        '404': { $ref: '#/components/responses/Error' }
        '429': { $ref: '#/components/responses/Error' }
  /services/{serviceID}:
    parameters: [ { $ref: '#/components/parameters/ServiceID' } ]
    get:
      summary: Get own service
      operationId: getOwnService
      responses:
        default: { $ref: '#/components/responses/Error' }
        '200': { description: Service, content: { application/json: { schema: { $ref: '#/components/schemas/Service' } } } }
        '404': { $ref: '#/components/responses/Error' }
    patch:
      summary: Update service
      operationId: updateService
      description: Full replacement of editable fields; status transitions DRAFT→ACTIVE/ARCHIVED, ACTIVE→ARCHIVED, ARCHIVED terminal.
      parameters: [ { $ref: '#/components/parameters/Csrf' } ]
      requestBody:
        required: true
        content: { application/json: { schema: { $ref: '#/components/schemas/ServiceUpdate' } } }
      responses:
        default: { $ref: '#/components/responses/Error' }
        '200': { description: Updated service, content: { application/json: { schema: { $ref: '#/components/schemas/Service' } } } }
        '400': { $ref: '#/components/responses/Error' }
        '404': { $ref: '#/components/responses/Error' }
        '429': { $ref: '#/components/responses/Error' }
  /providers/{providerID}/services/{serviceID}/passes:
    parameters: [ { $ref: '#/components/parameters/ProviderID' }, { $ref: '#/components/parameters/ServiceID' } ]
    post:
      summary: Create pass under own service
      operationId: createPass
      parameters: [ { $ref: '#/components/parameters/Csrf' } ]
      requestBody:
        required: true
        content: { application/json: { schema: { $ref: '#/components/schemas/PassInput' } } }
      responses:
        default: { $ref: '#/components/responses/Error' }
        '201': { description: Draft pass, content: { application/json: { schema: { $ref: '#/components/schemas/Pass' } } } }
        '400': { $ref: '#/components/responses/Error' }
        '404': { $ref: '#/components/responses/Error' }
        '429': { $ref: '#/components/responses/Error' }
...
  /catalog/passes/{passID}:
    parameters: [ { $ref: '#/components/parameters/PassID' } ]
    get:
      summary: Get own catalog pass
      operationId: getOwnPass
      responses:
        default: { $ref: '#/components/responses/Error' }
        '200': { description: Own catalog pass, content: { application/json: { schema: { $ref: '#/components/schemas/Pass' } } } }
        '404': { $ref: '#/components/responses/Error' }
    patch:
      summary: Update draft pass
      operationId: updatePass
      description: Full replacement of draft/unavailable editable fields; published passes are immutable.
      parameters: [ { $ref: '#/components/parameters/Csrf' } ]
      requestBody:
        required: true
        content: { application/json: { schema: { $ref: '#/components/schemas/PassInput' } } }
      responses:
        default: { $ref: '#/components/responses/Error' }
        '200': { description: Updated pass, content: { application/json: { schema: { $ref: '#/components/schemas/Pass' } } } }
        '400': { $ref: '#/components/responses/Error' }
        '404': { $ref: '#/components/responses/Error' }
        '409': { $ref: '#/components/responses/Error' }
        '429': { $ref: '#/components/responses/Error' }
    delete:
      summary: Delete (archive) a pass
      operationId: archivePass
      description: |
        Withdraws the pass from sale permanently. Only the provider that owns it
        may call this; any other identity receives 404. The pass row is retained
        with status ARCHIVED rather than deleted, because purchases, verified
        payments and already-purchased customer passes reference it: those
        records, their sessions and their redemption history are unaffected and
        remain readable. An archived pass disappears from the owner's catalog
        list and from all public discovery, cannot be purchased, edited or
        published, and archiving it again returns 409. Responds with the
        archived pass rather than 204.
      parameters: [ { $ref: '#/components/parameters/Csrf' } ]
      responses:
        default: { $ref: '#/components/responses/Error' }
        '200': { description: Archived pass, content: { application/json: { schema: { $ref: '#/components/schemas/Pass' } } } }
        '404': { $ref: '#/components/responses/Error' }
        '409': { $ref: '#/components/responses/Error' }
        '429': { $ref: '#/components/responses/Error' }
  /catalog/passes/{passID}/publish:
    parameters: [ { $ref: '#/components/parameters/PassID' } ]
    post:
      summary: Publish pass
      operationId: publishPass
      description: |
        Puts the pass on sale. Requires an active service, valid pass, and
        verified provider payout wallet. Accepts a pass in status DRAFT (never
        published) or UNAVAILABLE (withdrawn from the listing), so a withdrawn
        pass returns to sale on its own id, keeping its data and its purchase
        history. ACTIVE and ARCHIVED are both refused with 409: the pass is
        already on sale, or is permanently withdrawn.
      parameters: [ { $ref: '#/components/parameters/Csrf' } ]
      responses:
        default: { $ref: '#/components/responses/Error' }
        '200': { description: Published pass, content: { application/json: { schema: { $ref: '#/components/schemas/Pass' } } } }
        '404': { $ref: '#/components/responses/Error' }
        '409': { $ref: '#/components/responses/Error' }
        '429': { $ref: '#/components/responses/Error' }
  /catalog/passes/{passID}/unpublish:
    parameters: [ { $ref: '#/components/parameters/PassID' } ]
    post:
      summary: Remove a pass from the listing
      operationId: unpublishPass
      description: |
        Withdraws the pass from sale without deleting it, setting status
        UNAVAILABLE. This is not the delete transition: the pass keeps its id
        and all of its data, stays in the owner's catalog list so it can be
        edited, and can be put back on sale with POST
        /catalog/passes/{passID}/publish.

        While unpublished the pass is absent from public discovery, from the
        provider's public storefront and from the provider directory's pass
        count, has no public pass page, and cannot be purchased — a purchase
        attempt is refused with 409 PASS_UNAVAILABLE regardless of what the
        client had on screen.

        Nothing already sold is affected: purchased passes, their remaining
        sessions, purchase history, verified payments and redemption records
        are untouched and stay readable.

        Only the provider that owns the pass may call this; any other identity
        receives 404. Idempotent — unpublishing an already unpublished pass
        responds 200 with the unchanged pass and writes nothing. A DRAFT or
        ARCHIVED pass is refused with 409, because neither is on sale to
        withdraw.
      parameters: [ { $ref: '#/components/parameters/Csrf' } ]
      responses:
        default: { $ref: '#/components/responses/Error' }
        '200': { description: Unpublished pass, content: { application/json: { schema: { $ref: '#/components/schemas/Pass' } } } }
        '404': { $ref: '#/components/responses/Error' }
        '409': { $ref: '#/components/responses/Error' }
        '429': { $ref: '#/components/responses/Error' }
  /media:
    post:
      summary: Upload a pass cover image
      operationId: uploadMedia
      description: |
        Authenticated multipart upload of a JPEG or PNG pass cover. The original
        filename and declared content type are ignored. The server checks magic
        bytes, decodes the image, rejects oversized dimensions, re-encodes as
        JPEG and stores it under a generated UUID. Maximum upload size is 2 MiB.
      parameters: [ { $ref: '#/components/parameters/Csrf' } ]
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              additionalProperties: false
              required: [file]
              properties:
                file: { type: string, format: binary }
      responses:
        default: { $ref: '#/components/responses/Error' }
        '201': { description: Stored cover, content: { application/json: { schema: { $ref: '#/components/schemas/Media' } } } }
        '400': { $ref: '#/components/responses/Error' }
        '415': { $ref: '#/components/responses/Error' }
        '429': { $ref: '#/components/responses/Error' }
...
  /public/passes:
    get:
      summary: Discover published passes
      operationId: listPublicPasses
      description: |
        Up to 100 public passes, newest first with ID tie-breaker. Category is
        inherited from the service; omitted or empty means all categories
        including unclassified services. Unknown or repeated category values
        return VALIDATION_ERROR. `provider` narrows the list to one provider's
        storefront; a malformed or repeated value returns VALIDATION_ERROR, and
        an unknown provider yields an empty list, because absence and having
        published nothing are the same public fact.
      parameters:
        - { in: query, name: category, schema: { $ref: '#/components/schemas/Category' } }
        - { in: query, name: provider, schema: { type: string, format: uuid } }
      security: []
      responses:
        default: { $ref: '#/components/responses/Error' }
        '200': { description: 'Published, unexpired passes only', content: { application/json: { schema: { type: object, required: [items], properties: { items: { type: array, items: { $ref: '#/components/schemas/PublicPass' } } } } } } }
        '400': { $ref: '#/components/responses/Error' }
  /public/passes/{passID}:
    parameters: [ { $ref: '#/components/parameters/PassID' } ]
    get:
      summary: Get published pass
      operationId: getPublicPass
      security: []
      responses:
        default: { $ref: '#/components/responses/Error' }
        '200': { description: Published offer, content: { application/json: { schema: { $ref: '#/components/schemas/PublicPass' } } } }
        '404': { $ref: '#/components/responses/Error' }
  /public/providers:
    get:
      summary: List discoverable providers
      operationId: listPublicProviders
      description: |
        Up to 200 providers, ordered by how many passes each currently has on
        sale and then by name. A provider is listed when their payout wallet is
        verified and they have at least one pass that GET /public/passes would
        return, which is the same visibility rule the public catalog uses.
        passCount is that same count.
      security: []
      responses:
        default: { $ref: '#/components/responses/Error' }
        '200': { description: Discoverable providers, content: { application/json: { schema: { type: object, required: [items], properties: { items: { type: array, items: { $ref: '#/components/schemas/PublicProviderSummary' } } } } } } }
  /public/providers/{providerID}:
    parameters: [ { $ref: '#/components/parameters/ProviderID' } ]
    get:
      summary: Get public provider
      operationId: getPublicProvider
      security: []
      responses:
        default: { $ref: '#/components/responses/Error' }
        '200': { description: Public verified provider. Owner wallet is included for the identicon; payout wallet and identity id are not., content: { application/json: { schema: { $ref: '#/components/schemas/PublicProvider' } } } }
        '404': { $ref: '#/components/responses/Error' }
...
  /public/categories:
    get:
      summary: List canonical service category values
      operationId: listCategories
      security: []
      responses:
        default: { $ref: '#/components/responses/Error' }
        '200': { description: Controlled taxonomy; empty unclassified value is excluded, content: { application/json: { schema: { type: object, required: [items], properties: { items: { type: array, items: { type: string, enum: [fitness, tutoring, languages, coaching, wellness, music, beauty, consulting, mentoring] } } } } } } }
  /public/providers/by-slug/{slug}:
    get:
      summary: Get a public provider by stable slug
      operationId: getPublicProviderBySlug
      security: []
      parameters:
        - { in: path, name: slug, required: true, schema: { $ref: '#/components/schemas/ProviderSlug' } }
      responses:
        default: { $ref: '#/components/responses/Error' }
        '200': { description: Public verified provider. Owner wallet is included for the identicon; payout wallet and identity id are not., content: { application/json: { schema: { $ref: '#/components/schemas/PublicProvider' } } } }
        '400': { $ref: '#/components/responses/Error' }
        '404': { $ref: '#/components/responses/Error' }
...
components:
  securitySchemes:
    sessionCookie:
      type: apiKey
      in: cookie
      name: __Host-nimpass_session
      description: HTTPS production cookie; local HTTP development uses nimpass_session.
...
  responses:
    Error:
      description: Normalized error; codes include AUTH_REQUIRED, INVALID_SIGNATURE, CHALLENGE_EXPIRED, CHALLENGE_CONSUMED, FORBIDDEN, VALIDATION_ERROR, NOT_FOUND, CONFLICT, PAYMENT_CONFLICT, PASS_UNAVAILABLE, PASS_ALREADY_OWNED, PURCHASE_IN_SETTLEMENT, PASS_PURCHASE_CUTOFF, SELF_PURCHASE_NOT_ALLOWED, INTENT_EXPIRED, PASS_NOT_FOUND, PASS_NOT_OWNED, PASS_EXPIRED, PASS_COMPLETED, REDEMPTION_CHALLENGE_EXPIRED, REDEMPTION_ALREADY_CONSUMED, REDEMPTION_NOT_AUTHORIZED, INVALID_REDEMPTION_SIGNATURE, STALE_REDEMPTION_CHALLENGE, INVALID_REDEMPTION_TOKEN, RATE_LIMITED, CSRF_INVALID, ORIGIN_FORBIDDEN.
      content: { application/json: { schema: { $ref: '#/components/schemas/Error' } } }
  schemas:
...
    Provider:
      type: object
      required: [id, name, slug, headline, bio, avatarUrl, avatarVariant, location, payoutWallet, payoutVerifiedAt, createdAt, updatedAt]
      properties:
        id: { type: string, format: uuid }
        name: { type: string }
        slug: { $ref: '#/components/schemas/ProviderSlug' }
        headline: { type: string }
        bio: { type: string }
        avatarUrl: { type: string }
        avatarVariant: { type: integer, minimum: 0, maximum: 255 }
        location: { type: string }
        payoutWallet: { type: string, nullable: true }
        payoutVerifiedAt: { type: string, format: date-time, nullable: true }
        createdAt: { type: string, format: date-time }
        updatedAt: { type: string, format: date-time }
...